+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/ResponseParametersFrom.kt
@@ -15,6 +15,12 @@ sealed class ResponseParametersFrom {
         override val parameters: AuthenticationResponseParameters,
     ) : ResponseParametersFrom()
 
+    data class JweForJws(
+        val jweDecrypted: at.asitplus.signum.indispensable.josef.JweDecrypted<String>,
+        val parent: ResponseParametersFrom,
+        override val parameters: AuthenticationResponseParameters,
+    ) : ResponseParametersFrom()
+
     data class JweDecrypted(
         val jweDecrypted: at.asitplus.signum.indispensable.josef.JweDecrypted<AuthenticationResponseParameters>,
         val parent: ResponseParametersFrom,